Get Project Values

Generate a dataset of synthetic data for a REDCap project using the project design retrieved via the REDCap API.

Usage

get_project_values(
  redcap_uri,
  token,
  events = NULL,
  number_of_records_to_populate = 5
)

Arguments

redcap_uri

REDCap API URI

token

REDCap project API token

events

An optional vector of event names to fill (defaults to all events)

number_of_records_to_populate

An optional number of records to generate (default 5)

Value

A list of data frames (one per event, or just one for classic projects). Each list element contains a dataframe of synthetic data for each of the forms on that event. Each dataframe is ready to load into REDCap.